refactor(overlay): one file per pill, proven move-only (#2374 C1)
`OverlayLegacyViews.swift` held 18 declarations in 1,448 lines. It is deleted and
its contents relocated verbatim into thirteen files: ten leaf views under
`Overlay/Views/`, the two observable channels one level up under `Overlay/`
because they are presentation state rather than views, and the two rainbow brand
marks out of the overlay entirely into `App/Brand/`.

**This is NOT literally a no-op, and the exception is exactly two access keywords.**
`OverlayCapsuleBackground` and `DistressCapsuleBackground` were `private`. File-private
access does not survive a file split, so both widen to `internal`. The first is
unavoidable — eight views across eight new files construct it. The second is a
CHOICE: it could have stayed `private` beside `NotificationOverlayView`, at the cost
of splitting the shared background family whose literal counts the freeze guard
depends on. Nothing else in the moved source changed.

## The move is proven mechanically, not asserted

A diff of 1,448 moved lines is unreadable, which is the whole risk: a behavioural
change riding inside a wall of moved lines is invisible to a human reviewer.
`scripts/verify-view-split.py` reconstructs the original by concatenating the
thirteen destinations in declaration order and diffs against the base revision.

    MOVE-ONLY: 1320 lines reconstructed exactly from 13 files.
    residual_changed_lines: 0
    Allowed widenings applied: ['OverlayCapsuleBackground', 'DistressCapsuleBackground']
    Declared deletion: 10 comment line(s) at 38-47

The only permitted differences were enumerated BEFORE the move: the #2292 header
narration block, and the two access tokens above. A third would have been a finding.

Both the verifier and its 13-arm two-way control ship in this commit — a reviewer
asked to trust a MOVE-ONLY verdict from a script they cannot read is being handed
an unchecked premise. `scripts/*` is gitignored, so each needed an explicit
`.gitignore` exception or it would silently never ship (#2157).

## The freeze guard was migrated, not repaired

`CapsuleBackgroundFreezeTests` read the subject file by hardcoded path, and all
three of its tests broke differently under the split. Each break invites a
one-line repair that turns the suite green while deleting the coverage it exists
to provide. Every expectation is therefore UNCHANGED — 2, 1, 1; `reads >= 2`;
`count >= 8`. Only the source it reads changed:

- it now reads an EXPLICIT two-file set, `OverlayCapsuleBackgrounds.swift` plus
  `RecordingOverlayView.swift`, because the three frozen literals did not travel
  together — the notice-text literal has always lived inside `RecordingOverlayView`.
  Explicit rather than a directory glob: a glob would let an unrelated future
  declaration satisfy a frozen count after the intended literal disappeared, which
  is the same vacuity this suite was rewritten to close.
- it fails closed if either named file is missing or empty.
- the `hasPrefix("private struct DistressCapsuleBackground")` sentinel is replaced by
  a balanced brace walk. The widening above DELETED that exact string, and the
  resulting failure would have read as the struct being gone rather than as one
  keyword removed.

## What the build orchestrator changed, and what it did not

Codex reviewed this chunk as build orchestrator and returned five findings; round
two returned CHUNK-CLEAN. Three were adopted as given. Two were rejected with
measurement, and the measurements are the point:

- It asked for a declaration-level ownership check, because the aggregate count
  cannot tell which declaration holds the copy it found. The finding is real and
  is the suite's own documented weakness one level up — but it PREDATES this
  chunk, and this chunk NARROWED it: the counts ran over one 1,448-line file
  before and run over 655 lines now. Adding assertions to a chunk whose entire
  claim is "every expectation is unchanged" makes the one artifact a reviewer must
  read line by line harder to read. Filed as #2380 with a frozen mutation recipe
  and a `must_not_fire` row requiring the existing test to SURVIVE the mutant.
- It asked to replace the brace walk with `contains` over exact multi-line source
  text including indentation. That is whitespace-sensitive to a continuation line's
  exact indent — the comparison-narrower-than-the-language shape this repo keeps
  paying for. Measured instead: the struct region holds zero braces in comments,
  zero in string literals, no `#if`, no interpolation. Of the walk's four failure
  modes three fail loudly and one is silent, so the silent one is closed with a
  landing-line assertion and the loud ones are left alone. The residual
  textual-versus-lexical hazard is recorded as #2380's second frozen row.

  **One premise in that rejection was wrong and is corrected here rather than left
  standing.** It cited `swift-format` running in CI, inherited from the verifier's
  own docstring. No workflow mentions it. What exists is
  `.claude/scripts/swift-format-hook.sh`, a PostToolUse hook on `Edit|Write` that
  runs `swift-format format --in-place`. The conclusion survives and is stronger —
  a hook that REWRITES applies more pressure than a check that complains — and the
  docstring that carried the false claim is fixed in this commit.

  **That hook is also why every file here was written through Bash rather than the
  Edit/Write tools.** It fires after a Swift write and would have reformatted the
  moved code between the write and the equivalence check. Measured after the fact:
  all thirteen destinations and the migrated test are already format-clean (zero
  differing lines under `swift-format format`), so the hook would have been a no-op
  — but that is luck about this particular source, not a property of the method.

## Receipts

- Debug lane: 6596 tests, `** TEST SUCCEEDED **`, exit 0, zero
  `^✘ Test ".*" failed after` marks.
- Release lane: 5995 tests, exit 0. The 601-test gap to Debug is structural —
  `#if DEBUG` tests are absent from the Release binary by construction.
- All three freeze tests pass with every expectation unchanged: 2, 1, 1;
  `reads >= 2`; `count >= 8`.
- `verify-view-split.test.sh`: 13 arms, three distinct exit codes, all passed.
  Arms 11 and 12 are new here and are a PAIR: adding a blank line to a destination is
  MOVE-ONLY, indenting one nonblank line is CONTENT CHANGED. They exist because the
  verifier's docstring said it was "blind to whitespace-only changes", which
  overclaims in the dangerous direction — a reader would conclude a reindent of moved
  code passes unreported, and it does not. Found by review round three; the claim is
  now controlled rather than asserted.
- Import manifest: all 13 destinations carry the source file's original four
  imports, unminimised, 0 deviations.
- `tuist generate` from scratch: all 13 files present in the project, 0 references
  to the deleted file.

Part of #2374

Self-audit-grep: grep -c PreviewPillPalette. Sources/EnviousWisprAppKit/App/Overlay/Views/OverlayCapsuleBackgrounds.swift
Self-audit-owner: Tests/EnviousWisprTests/App/CapsuleBackgroundFreezeTests.swift
